requirements.txt
fastapi
uvicorn
langchain
langserve
python-dotenv
openai
langchain-openai
langchain-community
